backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
opencv-python-headless>=4.8.0
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0
aiofiles>=23.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
yt-dlp>=2023.0.0
python-dotenv>=1.0.0
